+++ src/main/java/com/project/poblog/controller/user/UserController.java
@@ -1,9 +1,9 @@
 package com.project.poblog.controller.user;
 
-import com.project.poblog.dto.user.request.JoinReq;
-import com.project.poblog.dto.user.response.JoinRes;
+import com.project.poblog.domain.user.dto.request.JoinReq;
+import com.project.poblog.domain.user.dto.response.JoinRes;
 import com.project.poblog.global.response.CommonResponse;
-import com.project.poblog.service.user.UserService;
+import com.project.poblog.domain.user.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
