+++ src/main/java/com/project/poblog/domain/user/dto/UserMapper.java
@@ -1,0 +1,18 @@
+package com.project.poblog.domain.user.dto;
+
+import com.project.poblog.domain.user.entity.User;
+import com.project.poblog.domain.user.dto.request.JoinReq;
+import com.project.poblog.domain.user.dto.response.LoginRes;
+import com.project.poblog.domain.user.dto.response.JoinRes;
+import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+
+@Mapper(componentModel = "spring") //스프링 컨텍스트 등록
+public interface UserMapper {
+
+    @Mapping(target = "role", expression = "java(com.project.poblog.domain.user.entity.Role.MEMBER)")
+    User toEntity(JoinReq request);
+
+    JoinRes toRegisterResponse(User user);
+    LoginRes toLoginUserResponse(User user);
+}
