--- src/main/java/com/project/poblog/dto/user/response/LoginRes.java
@@ -1,12 +1,0 @@
-package com.project.poblog.dto.user.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class LoginRes {
-    private String nickname;
-}
